Don't rebuild ttf-ms-win10 if already built

+++ build/make-iso.sh
@@ -66,12 +66,17 @@ function image_usb() {
 function build_win10_fonts() {
     local PKG_PATH="$1"
 
-    mkdir -p /var/tmp/win10-fonts
-    $BASE_DIR/misc/build-ttf-ms-win10.sh "/var/tmp/win10-fonts"
+    if ! find "$PKG_PATH" -name "*ttf-ms-win10*" &> /dev/null; then
+        mkdir -p /var/tmp/win10-fonts
+        $BASE_DIR/misc/build-ttf-ms-win10.sh "/var/tmp/win10-fonts"
 
-    sudo mv /var/tmp/win10-fonts/* "$PKG_PATH"
+        sudo mv /var/tmp/win10-fonts/* "$PKG_PATH"
 
-    rm -r /var/tmp/win10-fonts
+        rm -r /var/tmp/win10-fonts
+    else
+        echo "ttf-ms-win10 package already found in $PKG_PATH"
+        echo "Not going to rebuild tff-ms-win10"{
+    fi
 }
 
 function build_aur_packages() {
